Remove unused and no-longer-supported gAtomService
Components.interfaces.nsIAtomService is gone in recent Thunderbird
versions. That's OK, because we're no longer actually using it; we
just forgot to remove the line of code that attempts to fetch it. This
non-functional change removes that line of code to fix an
incompatibility with recent Thunderbird versions (57+ or 58+, not
certain which).